requirements.txt
numpy
panda
scikit_learn
torch
torchvision
torchaudio
